# docs/openapi.yaml
swagger: "2.0"
info:
  title: Sample API
  description: API description in Markdown.
  version: 1.0.0
paths:
...
  /transactions/create:
    post:
      summary: Create transaction
      description: ""
      operationId: createTransaction
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          description: The transaction object that needs to be added to the system
          required: true
          schema:
            $ref: "#/definitions/Transaction"
      responses:
        "200":
          description: successful operation
          schema:
            type: string
        "429":
          description: Too many requests
      security:
        - jwt: [ ]
...
definitions:
...
  Transaction:
    type: object
    properties:
      amount:
        type: number
      status:
        type: string
...
securityDefinitions:
  jwt:
    type: apiKey
    name: Authorization
    in: header
    description: "JWT token must be provided in format 'Bearer <token>'."